// apps/api/src/lib/types.ts
export interface UserDoc {
  id: string;
  email: string;
  name: string;
  avatarUrl?: string;
  createdAt: number;
  lastLogin: number;
}